fix: extract lua code fix: move test class to the test folder fix  extract classes to their own file
