+++ lib/throttle_machines/hedged_request.rb
@@ -105,26 +105,6 @@ module ThrottleMachines
     end
   end
 
-  # Hedged request with circuit breaker integration
-  class HedgedBreaker
-    def initialize(breakers, delay: 0.05)
-      @breakers = Array(breakers)
-      @hedged = HedgedRequest.new(
-        delay: delay,
-        max_attempts: @breakers.size
-      )
-    end
-
-    def run(&)
-      @hedged.run do |attempt|
-        breaker = @breakers[attempt]
-        next if breaker.nil?
-
-        breaker.call(&)
-      end
-    end
-  end
-
   # Convenience method
   def self.hedged_request(**, &)
     hedged = HedgedRequest.new(**)
