+++ lib/throttle_machines/storage/memory.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require_relative 'base'
 require 'concurrent'
 
 module ThrottleMachines
