+++ lib/throttle_machines/storage/redis/peek_token_bucket.lua
@@ -1,0 +1,18 @@
+local key = KEYS[1]
+local capacity = tonumber(ARGV[1])
+local refill_rate = tonumber(ARGV[2])
+local now = tonumber(ARGV[3])
+
+local bucket = redis.call('HMGET', key, 'tokens', 'last_refill')
+local tokens = tonumber(bucket[1]) or capacity
+local last_refill = tonumber(bucket[2]) or now
+
+-- Calculate tokens without modifying
+local elapsed = now - last_refill
+local tokens_to_add = elapsed * refill_rate
+tokens = math.min(tokens + tokens_to_add, capacity)
+
+local allow = tokens >= 1
+local tokens_after = allow and (tokens - 1) or 0
+
+return { allow and 1 or 0, tokens_after }
